Move TextileNode to Services
