+++ App/Sagas/NodeLifecycle.ts
@@ -10,7 +10,7 @@ import RNPushNotification from 'react-native-push-notification'
 import StorageActions from '../Redux/StorageRedux'
 import TextileNodeActions from '../Redux/TextileNodeRedux'
 import { PreferencesSelectors } from '../Redux/PreferencesRedux'
-import TextileNode from '../../TextileNode'
+import TextileNode from '../Services/TextileNode'
 import { RootAction } from '../Redux/Types'
 import {Threads, ThreadName} from '../Models/TextileTypes'
 import {logNewEvent} from './DeviceLogs'
