+++ App/Sagas/PhotoViewingSagas.ts
@@ -5,7 +5,7 @@ import { ActionType, getType } from 'typesafe-actions'
 import PhotoViewingActions from '../Redux/PhotoViewingRedux'
 import UIActions from '../Redux/UIRedux'
 import { photoAndComment, shouldNavigateToNewThread, shouldSelectNewThread, photoToShareToNewThread } from '../Redux/PhotoViewingSelectors'
-import TextileNode from '../../TextileNode'
+import TextileNode from '../Services/TextileNode'
 import { Threads, Photo, BlockId, PhotoId } from '../Models/TextileTypes'
 import NavigationService from '../Services/NavigationService'
 import { shareWalletImage } from './ImageSharingSagas'
