+++ App/Sagas/TextileSagas.ts
@@ -15,7 +15,7 @@ import { call, put, select, take, fork } from 'redux-saga/effects'
 import RNFS from 'react-native-fs'
 import Config from 'react-native-config'
 import NavigationService from '../Services/NavigationService'
-import TextileNode from '../../TextileNode'
+import TextileNode from '../Services/TextileNode'
 import { getPhotos } from '../Services/CameraRoll'
 import * as NotificationsSagas from './NotificationsSagas'
 import StartupActions from '../Redux/StartupRedux'
