+++ App/Sagas/ThreadsSagas.ts
@@ -14,7 +14,7 @@ import { call, put, select, fork } from 'redux-saga/effects'
 import ThreadsActions, { ThreadsSelectors } from '../Redux/ThreadsRedux'
 import { ActionType } from 'typesafe-actions'
 import { BlockId, ExternalInvite, Thread, Threads, ThreadId } from '../Models/TextileTypes'
-import TextileNode from '../../TextileNode'
+import TextileNode from '../Services/TextileNode'
 import DeepLink from '../Services/DeepLink'
 import PhotoViewingActions from '../Redux/PhotoViewingRedux'
 import NavigationService from '../Services/NavigationService'
